+++ src/components/animations/text.tsx
